Move curated font names to fonts package
